feat(suite): move autostart setting to application tab
