+++ packages/suite/src/views/settings/SettingsDebug/SettingsDebug.tsx
@@ -20,9 +20,7 @@ import { Backends } from './Backends';
 import { selectSuiteFlags } from 'src/reducers/suite/suiteReducer';
 import { useSelector } from 'src/hooks/suite';
 import { PreField } from './PreField';
-import { AutoStart } from './AutoStart';
 import { Tor } from './Tor';
-import { ShowOnTray } from './ShowOnTray';
 
 export const SettingsDebug = () => {
     const flags = useSelector(selectSuiteFlags);
@@ -56,12 +54,6 @@ export const SettingsDebug = () => {
             <SettingsSection title="Testing">
                 <ThrowTestingError />
             </SettingsSection>
-            {!isWeb() && (
-                <SettingsSection title="Application">
-                    <AutoStart />
-                    <ShowOnTray />
-                </SettingsSection>
-            )}
             {!isWeb() && (
                 <SettingsSection title="Transport backends">
                     <TransportBackends />
