+++ packages/suite/src/views/settings/SettingsGeneral/AutoStart.tsx
