+++ packages/suite/src/views/settings/SettingsGeneral/ShowOnTray.tsx
